EDI Transmission Processing Started
Added Installer, Processor, ProcessThread and ProcessTask classes to the
package `com.is2300.jedi.edi`.

Made small changes to the interface of the `EDISettingsPanel`. I removed
the "Database Name" field, because the name is going to be set by us and
not by our clients. I split the Database Server Path and Port into two
separate fields, using a JFormattedTextField with the number formatter
factory, which is custom and only allows a 4 digit number, with no 
punctuation to be entered into the Port field.