MySQL Driver Added
Added the MySQL Jarchive to the Release Files so that we will have the
appropriate driver for connecting to the MySQL database server.

Set up the installer class to execute a `java.util.Timer` according to
the time frame established by the user in the settings file. Each time
the `Timer` executes, it runs a `java.util.TimerTask`, which in turn
spawns a `java.lang.Thread` as a daemon to handle all of the 
processing.

The `Thread` default constructor takes no arguments, initializes the
`Thread's` member fields, then calls the `shepherd()` method that takes
care of shepherding the processing functions.