File Parsing Begun
Work on the file parsing routines has begun. So far, the file gets
opened, the lines are counted due to the file being read into a list as
lines, the lines get split into their individual fields and stored into
the appropriate lists, the Functional Groups are being validated.

As yet, no data is being retained and we are not testing the database
connectivity. Will probably start on that in a couple days.