Working on File Parsing
Have gotten the file parsing process to write to the Output Window and
am now building a report message in tandem with writing to the Output
Window. The report message then gets automatically saved once all 
processing has completed, to the Output Path set in the Options dialog.
If there is no setting stored, the default output path is the user's
home folder. The report file is named as: YYYY-MM-DD.HH-MM-SS.rpt

The last method call in the `shepherd()` method remains the call to 
`selfDestruct()` for preparing all objects for garbage collection. The
next to last call in `shepherd()` is to `saveReport()` to save the 
processing report to file.

### Next Up
Going to begin working on adding the data from the incoming EDI file to
the database for audit purposes. This involves writing various info to
the following three (3) database tables:

* is_edi_audits
* is_edi_audit_grp_details
* is_edi_audit_doc_details