MySQL Data Storage Commenced
Started working on storing data from the incoming EDI transmission file
to the MySQL database, is_jedi, and putting information into the audit
tables.

Have one issue in the code being committed at this time: date math is
not taking place as one would expect. I'm displaying information during
the processing cycle to the Output window of the Platform and showing a
summary report at the end of the output that looks like this:

     Sat: 03/11/2017 - 09:46:21: Initializing EDI processor...
     ...
     Sat: 03/11/2017 - 09:46:31:  Parsing Complete.
	    Envelopes:  61
	       Groups:  61
     Transactions:  124
     Processed in 0.000000 minute(s)

As you can see in the above example, the date math is not correct for
how long the processing took. The process took a total of ten (10) or
so seconds, but the date math shows "0.000000" for the processing time.
I am unable, so far, to figure out why this is. I've gone from using 
the `java.util.Calendar.compare()` method to doing the math by hand by
extracting the start and end time milliseconds by using the
`java.util.Calendar.getTimeInMillis()` function, then subtracting, and
then doing the division. However, no matter how I do this, the result
is the same: 0.000000. I've used both a `float` and a `double` for
performing the division into and both come out the same.

If anyone has any ideas on this, please comment on this commit and let
me know. We have got to solve this weird issue before we get to a 
release, otherwise, we need to drop the date math portion of the 
process summary, which I really don't want to do.

***Just a note:** I have posted this issue to a Java mailing list, so
hopefully, we'll get some help that way.*