Continuing Work on Parsing
I have fixed the date math issue from the last commit involving showing
the elapsed processing time. I changed the code from this:

        // Output the lapsed time of the document processing.
        this.end = Calendar.getInstance();
        long strt = this.start.getTimeInMillis();
        long fnsh = this.end.getTimeInMillis();
        long mins = fnsh - strt;
        float procTime = mins / 60000;

...to this:

        // Output the lapsed time of the document processing.
        this.end = Calendar.getInstance();
        long strt = this.start.getTimeInMillis();
        long fnsh = this.end.getTimeInMillis();
        long mins = fnsh - strt;
        float procTime = mins / 60000f;

That little `f` after the `60000` made all the difference in the world.
Now, our elapsed processing time that is printed to the Output Window
is showing an actual fraction of a minute. This gives me hope that it
will show the correct amount of time from this point forward.

I've also gotten the interchange envelope validation running and it is
working as expected.

Work continues and we keep moving closer to something that actually has
some functionality to it...