Continuing Work on Parsing
Have added the validation of Invoice (810) Transaction Set segments, 
but am having trouble finding an EDI sample file with an 810 
transaction set in it. Have 2 years worth of EDI files, but have to
reformat them, as they were processed by EDIMapper (VB6 project) and
have the `\r` line feeds replaced with `]` right square brackets. Need
to reconfigure those files to put the `\r` line feeds back in.

Work continues and we keep moving closer to something that actually has
some functionality to it...