Continuing Work on Parsing
Created a generic SegmentValidator class that takes four (4) parameters
and validates any EDI segment. The first parameter is the segment to be
validated, the last parameter is a list of validation rules to be used
in validating the segment fields.

This functionality is currently throwing a NullPointerException from 
the `com.is2300.jedi.edi.utils.Utils` class' `string2Time()` function.

This functionality still needs to be tested, but I am pressed for time
right now. Will get back to this on Thursday morning (3/16/2017).