Continuing Work on Parsing
Tested the `SegmentValidator` class and it is now working. I was getting
`NullPointerException` errors and spent about two (2) hours debugging
before I figured out the logic error in the code. My problem is that I
was concentrating on figuring out what was going wrong in the new class,
`SegmentValidator` because I have never done such generic programming
before. However, the problem was not in `SegmentValidator`. The problem,
as it turns out, was in my `switch() case:` block in the class,
`Validate824Segments`. I had not put the appropriate `break` statement
in each of the `case:` blocks, so the `NullPointerException` errors
were due to attempting validation of a segment with the validation 
rules of the prior segment(s). I inserted the `break` statements and
everything now works as expected.

Processing elapsed time is now up to about nine (9) seconds for a file
that has almost 4,000 lines. I will start testing with larger incoming
EDI transmission files...I have a couple that are over 13,000 lines in
each one.

#### Quick Note
I built a quick and dirty J2SE/Swing application for repairing EDI
transmission files. The repair is simply replacing the `]` character
with the line feed character `\r`. The line feeds were stripped out by
the old EDIMapper program because VB6 did not play well with `\r` line
feed characters.

Anyway, I have set it to be ignored by `git`, but if you would like to
have it, let me know and I will email you the JAR file.